updated card_file_manager.py to create a directory for audio files if there wasn't one already. It also creates an audio directory config in the config file to be used later on.
implemented text to speech using a japanese kokoro tts model.

flashcard objects were changed to have 2 different filepath variables. one for the word and one
for the sentences.

main.py was updated to include the implemented functionality of text to speech. cards now have
text to speech audio generated for them.

README updated to include instructions for downloading dependencies for the tts model.